(The last update was accidentally untagged.)
This is a centralized rewrite of Wakaba. Along with  bug fixes, many files that were once necessary for each board (including wakaba.pl itself) can now be placed in a single directory. Others, such as the CSS files, rules.html, and the local board configuration, board_config.pl, are placed inside each board directory to provide local board management options. The directory structure has been altered to show this. To install a new board, drag in the contents of "board" to a new directory and do <PATH_TO_SCRIPT>/wakaba.pl?task=admin&board=<PATH_TO_NEW_DIR_IN_RELATION_TO_SCRIPT> in the web browser. Then, rebuild cache.

Of particular note in the source is the new Board module, as well as some new configuration options and loads of reorganizing the code.

This is being used on a trial basis on Desuchan. Currently, /temp/, /nipa/, and /o/ are running on this centralized Wakaba.

This latest update carries over Schmidt's mod of burichan.css, which wasn't present on the repo until now. 

git-svn-id: https://suigintou.weedy.ca/svn/desuchan/trunk@27 4871d5c7-3bdd-4872-8bc4-3d792a1ad158